Neo4j Server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
llama-index>=0.8.11
python-dotenv>=1.0.0
neo4j>=5.8.1
# add any other dependencies you need
